core/requirements.txt
cassandra-driver==3.14.0
elasticsearch==7.8.0
confluent-kafka==1.2.0
urllib3
requests
